change wallet column type to bigInteger

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -24,7 +24,7 @@ class CreateUsersTable extends Migration
 			$table->string('avatar')->nullable();
 			$table->boolean('is_active')->default(1);
             $table->enum('role',['user','company','pro','admin'])->default('user');
-            $table->float('wallet')->default(0);
+            $table->bigInteger('wallet')->default(0);
             $table->rememberToken();
             $table->timestamps();
         });
